backend/requirements.txt
# ---Core runtime---
requests>=2.31.0
python-dotenv>=1.0.0

# ---FastAPI (Issue #27: Cloud Run Deployment)---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.0

# LangChain / LangGraph stack ---
langchain-core>=0.3.64
langchain-openai>=0.2.0,<0.3.0
langgraph>=0.4.0
langgraph-sdk>=0.2.0
langgraph-cli>=0.1.0
langgraph-api>=0.0.1
langsmith>=0.3.45

# ---Compatibility helpers---
typing-extensions>=4.8.0
backports.zoneinfo; python_version < "3.9"

# ---Gmail API (Issue #6)---
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# ---PDF Processing (Issue #17)---
pypdf>=3.17.0
langchain-community>=0.3.0

# ---Supabase (Issue #24)---
supabase>=2.22.0

# ---Auth (Supabase Auth統合)---
python-jose[cryptography]>=3.3.0

# ---Testing---
pytest>=8.0.0
pytest-asyncio>=1.3.0
pytest-mock>=3.14.0
pytest-cov>=7.0.0

# ---動画生成（Video Narration Feature）---
moviepy>=1.0.3         # 動画生成（画像+音声合成）
pillow>=10.0.0         # 画像処理
playwright>=1.49.0     # HTML→PNG変換（スライド画像生成）

# ---Cloud Tasks（動画生成ジョブ投入）---
google-cloud-tasks>=2.16.0